+++ example/lib/choose.dart
@@ -1,7 +1,9 @@
 import 'package:flutter/material.dart';
+import 'package:openidconnect_example/client.dart';
 import 'device.dart';
 import 'interactive.dart';
 import 'password.dart';
+import 'redirectloop_result_page.dart';
 
 class ChoosePage extends StatelessWidget {
   @override
@@ -36,6 +38,22 @@ class ChoosePage extends StatelessWidget {
                 ),
               ),
               child: Text("Password"),
+            ),
+            TextButton(
+              onPressed: () => Navigator.of(context).push(
+                MaterialPageRoute(
+                  builder: (context) => ClientPage(),
+                ),
+              ),
+              child: Text("Client Usage Example"),
+            ),
+            TextButton(
+              onPressed: () => Navigator.of(context).push(
+                MaterialPageRoute(
+                  builder: (context) => RedirectLoopResultPage(),
+                ),
+              ),
+              child: Text("Redirect Loop Result Page"),
             ),
           ],
         ),