+++ example/lib/client.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/material.dart';
+
+class ClientPage extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    throw UnimplementedError();
+  }
+}