+++ example/lib/credentials.dart
@@ -1,0 +1,13 @@
+final defaultDiscoveryUrl =
+    "https://yourdomain.com/.well-known/openid-configuration";
+final defaultClientId = "Admin";
+final String? defaultClientSecret = "setme";
+final String? defaultAudience = "api";
+final defaultRedirectUrl = "http://localhost:50335/callback.html";
+final defaultscopes = [
+  "openid",
+  "profile",
+  "email",
+  "address",
+  "offline_access"
+];