+++ openidconnect/lib/src/models/requests/refresh_request.dart
@@ -4,7 +4,6 @@ class RefreshRequest extends TokenRequest {
   RefreshRequest({
     required String clientId,
     String? clientSecret,
-    required String redirectUrl,
     required Iterable<String> scopes,
     required String refreshToken,
     required OpenIdConfiguration configuration,