+++ openidconnect_platform_interface/lib/openidconnect_platform_interface.dart
@@ -29,11 +29,14 @@ abstract class OpenIdConnectPlatform extends PlatformInterface {
     _instance = instance;
   }
 
-  Future<String> authorizeInteractive({
+  Future<String?> authorizeInteractive({
     required String title,
     required String authorizationUrl,
     required String redirectUrl,
     required int popupWidth,
     required int popupHeight,
+    bool useWebRedirectLoop = false,
   });
+
+  Future<String?> processStartup();
 }