+++ openidconnect_platform_interface/lib/src/exceptions/authentication_exception.dart
@@ -1,5 +1,0 @@
-part of openidconnect_platform_interface;
-
-class AuthenticationException extends OpenIdConnectException {
-  AuthenticationException([String? errorMessage]) : super(errorMessage);
-}