+++ openidconnect_platform_interface/lib/src/exceptions/http_response_exception.dart
@@ -1,9 +1,0 @@
-part of openidconnect_platform_interface;
-
-class HttpResponseException implements Exception {
-  final String? errorMessage;
-  HttpResponseException([this.errorMessage]);
-
-  @override
-  String toString() => errorMessage ?? "Unknown";
-}