+++ openidconnect_platform_interface/lib/src/exceptions/logout_exception.dart
@@ -1,5 +1,0 @@
-part of openidconnect_platform_interface;
-
-class LogoutException extends OpenIdConnectException {
-  LogoutException([String? errorMessage]) : super(errorMessage);
-}