+++ openidconnect_platform_interface/lib/src/exceptions/openidconnect_exception.dart
@@ -1,9 +1,0 @@
-part of openidconnect_platform_interface;
-
-abstract class OpenIdConnectException implements Exception {
-  final String? errorMessage;
-  OpenIdConnectException([this.errorMessage]);
-
-  @override
-  String toString() => errorMessage ?? "Unknown";
-}