+++ openidconnect_platform_interface/lib/src/exceptions/revoke_exception.dart
@@ -1,5 +1,0 @@
-part of openidconnect_platform_interface;
-
-class RevokeException extends OpenIdConnectException {
-  RevokeException([String? errorMessage]) : super(errorMessage);
-}