+++ openidconnect_platform_interface/lib/src/exceptions/user_info_exception.dart
@@ -1,5 +1,0 @@
-part of openidconnect_platform_interface;
-
-class UserInfoException extends OpenIdConnectException {
-  UserInfoException([String? errorMessage]) : super(errorMessage);
-}