+++ openidconnect/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.8
+
+- Change how the client library is instantiated using the static future: OpenIdConnectClient.create to ensure that everything is initialized. This will prepare for handling web logins that use a redirect loop instead of a popup window and any other platform specific initialization code as necessary.
+- Enables interactive login processStartup call for the interface.
+
 ## 1.0.7+1
 
 - Bug fix