+++ openidconnect_platform_interface/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.0.2
+
+- Expanded the response for processStartup
+- Add parameters for if a web login redirect loop should be used
+
+## 1.0.1
+
+- Add processStartup to handle web redirect loop.
+
 ## 1.0.0
 
 - Initial Release