+++ openidconnect_web/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.9
+
+- Implement redirect loop functionality in addition to the default web popup.
+
 ## 1.0.7
 
 - Fix readme instructions